Moved content type model classes to correct namespace, usings cleanup, related files update.
