+++ src/content/Kentico.Kontent.Boilerplate/Models/ContentTypes/Brewer.cs
@@ -1,8 +1,4 @@
-using System;
-using System.Collections.Generic;
-using Kentico.Kontent.Delivery.Abstractions;
-
-namespace Kentico.Kontent.Boilerplate.Models
+namespace Kentico.Kontent.Boilerplate.Models.ContentTypes
 {
     public partial class Brewer
     {
