+++ src/content/Kentico.Kontent.Boilerplate/Views/Home/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<Article>
+﻿@model IEnumerable<Kentico.Kontent.Boilerplate.Models.ContentTypes.Article>
 @{
     ViewData["Title"] = "Home Page";
 }
