+++ src/content/Kentico.Kontent.Boilerplate/Views/Shared/DisplayTemplates/Article.cshtml
@@ -1,4 +1,4 @@
-﻿@model Article
+﻿@model Kentico.Kontent.Boilerplate.Models.ContentTypes.Article
 
 @{
     Layout = null;
